ReturnInfo as interface #1053
